requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gradio
pydantic
groq
serpapi
